Refactoring Changes.
Updated Board class so that moves can only be made if game isn't in an over state.  Added helper methods to return different cells of board.
New Cell Enum used for controlled and descriptive way of making moves and also for showing possible moves in menu.
Added Constants for Play a new game Functionality.  Passing gameOver flag to Move derived classes used when seeing if a move is valid.  Using the factory pattern to control the creation of players so that only exactly two players are ever created, when creating a game.  Added test cases to test for the possible win outcomes as well as a tied game.